+++ packages/jshark/src/JShark/Compiler/CompileTerminal.hs
@@ -16,6 +16,8 @@ module JShark.Compiler.CompileTerminal
   , renderSingleDone
   , renderStatsTable
   , styled
+  , padLeft
+  , padRight
   )
 where