+++ packages/jshark/src/JShark/Compiler/Flat.hs
@@ -27,6 +27,7 @@ module JShark.Compiler.Flat
   , encodeFlatNode
   , emptySoaSideAcc
   , flatNodeIsEffect
+  , irNodeIsEffect
   , flatNodeChildRefs
   , flatArgRef
   , flatFieldRef